Split humanreadble.go into tablegenerator.go and tableprinter.go
