+++ pkg/printers/BUILD
@@ -9,8 +9,9 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
-        "humanreadable.go",
         "interface.go",
+        "tablegenerator.go",
+        "tableprinter.go",
         "tabwriter.go",
     ],
     importpath = "k8s.io/kubernetes/pkg/printers",
@@ -45,10 +46,10 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = ["humanreadable_test.go"],
+    srcs = ["tableprinter_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/apis/core:go_default_library",
+        "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1beta1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/runtime:go_default_library",
